give canvas size on menu bar an icon, refactor packages slightly
