+++ src/GUI/MenuBarListener.java
@@ -1,4 +1,4 @@
-package Toolstrip;
+package GUI;
 
 public interface MenuBarListener {
     void onFileOpened(String filePath);
